Address clippy issues.
In particular the "module inception" issue where we had some modules in the
pattern

mod foo;
pub use foo::*;

under the foo.
