+++ rust-src/concordium_base/src/curve_arithmetic/bls12_381_instance.rs
@@ -1,4 +1,4 @@
-use super::{bls12_381_g1hash::*, bls12_381_g2hash::*, curve_arithmetic::*};
+use super::{bls12_381_g1hash::*, bls12_381_g2hash::*, *};
 use byteorder::ReadBytesExt;
 use ff::{Field, PrimeField};
 use group::{CurveAffine, CurveProjective, EncodedPoint};
