+++ rust-src/concordium_base/src/eddsa_ed25519/ffi.rs
